requirements.txt
Django==2.1.2
django-ckeditor==5.6.1
django-cleanup==3.0.1
django-crispy-forms==1.7.2
django-js-asset==1.1.0




